soe work on console
finding a Python library that provides non-blocking
reads from the terminal in MacOS is a great big PITA
just like every other attempt to do something in Python.
Half the answers are "you should not want to do that".
Another half have "simple" solutions of amaing complexity.
And A third half don't work.